// src/app/model/IProcuracion.ts
export class IListadoSiniestro {
  idsiniestro: string="";
  poliza: string="";
  placa: string="";
  fechaini: string="";
  fechafin: string="";
  nombre: string="";
  NroDePagina: string="";
  RegPorPag: string="";
  idsemaforo: string="";
  idperfil: string="";
  usuario: string="";
  icodigorenovacion: string=""; 
  idordenatencion: string="";
}

export class IProcuracion {
    idsiniestro: string="";
    hechosSiniestros: string="";
    apreciacionSiniestro: string="";
    dFecReporteProc: string="";
    dHoraReporteProc: string="";
    dFecInicioProc: string="";
    dHoraInicioProc: string="";
    dFecFinProc: string="";
    dHoraFinProc: string="";
    tipoasesoria: string="";
    ikilometraje: string="";
    fechaAsignacion: string="";
    horaAsignacion: string="";
    procurador_id: string="";
}
  

export class IActualizarSiniestro {
  idsiniestro: string="";
  idpoliza: string=""; 
  icodigorenovacion: string=""; 
  idvehiculo: string=""; 
  smidciaseguros: string=""; 
  iestadosiniestro: string=""; 
  dFecNotificacion: string=""; 
  dHoraNotificacion: string="";
  dFecRegistro: string=""; 
  dHoraRegistro: string=""; 
  dFecModifi: string=""; 
  dHoraModifi: string="";
  idconsecuencia: string=""; 
  dFecOcurrencia: string=""; 
  dHoraOcurrencia: string=""; 
  vlugarsiniestro: string=""; 
  iocupantes: string=""; 
  vtelef_declarante: string=""; 
  iparentaseg_declarante: string=""; 
  vmaildeclarante: string=""; 
  vlicencia: string=""; 
  iparentaseg_conductor: string=""; 
  vtelef_conductor: string=""; 
  vemail_conductor: string=""; 
  dvencilicencia: string=""; 
  idcomisaria: string=""; 
  vcategoria: string=""; 
  vdetasiniestro: string=""; 
  nidusuario: string=""; 
  idgravedad: string=""; 
  smcausa: string=""; 
  idetallecausa: string=""; 
  idtipo_decla: string=""; 
  idtipo_docid_decla: string=""; 
  idnrodoc_decla: string=""; 
  vnombres_decla: string=""; 
  vapellidopat_decla: string=""; 
  vapellidomat_decla: string=""; 
  idtipo_conduc: string=""; 
  idtipo_docid_conduc: string=""; 
  idnrodoc_conduc: string=""; 
  vnombres_conduc: string=""; 
  vapellidopat_conduc: string=""; 
  vapellidomat_conduc: string=""; 
  smidubigeo: string=""; 
  dFecDenunciaP: string=""; 
  dHoraDenunciaP: string=""; 
  dFecCambioEstado: string=""; 
  dUserAprobado: string=""; 
  smidanalista: string=""; 
  idaseguradora: string=""; 
  usuarioLogin: string=""; 
  idsemaforo: string=""; 
  usuario: string=""; 
  ip: string=""; 
}